More functions and a clearer path forward
I added some setters and getters for my board class. The most important of these is the get_score() function. This function allows for a board to be given a score based on material (this will need to be updated later to include scoring based on king safety, piece mobility/position, etc.)
The next major function I need to write is the solver, which calculates the AI's move and then updates the board accordingly. For this I will be using the minimax strategy.