(Some) success.
After some painful debugging, my AI seems to play a few moves logically enough. My baseline is pretty low. Basically, the AI seems to take free material when it is offered. The next step will be to add to the set_score function. This will allow the AI to choose moves based on mobility, control of the center, etc.

At this point, I would say that I've successfully achieved some of the functionality I desire, but I haven't everywhere adhered to RAII (!). That is the other thing I'd like to address soon.