Final touches
I ironed out a few hiccups, commented my functions and created a display message for in the case that the user checkmates the computer. At this point, I have a working version, though there is a lot that could be added: allowing the user to play black as well as white, having the user enter moves in algebraic notation, allow for castling and en passant.
Of course, this project was originally to create an AI to play the crazyhouse variant of chess. That would also be something I'd like to add in the future.